feat(transition): Remove react-transition-group
I'm not a big fan of the changes to conditionally provide the node if a
nodeRef has been passed in the latest version of react-transition-group.
Since I already have useTrasition and useCSSTransition hooks, I went
ahead and just removed react-transition-group as a dependency and added
the Transition and CSSTransition components.

Interestingly enough, these new hooks and components actually have
better behavior than react-transition-group and I managed to fix a few
slight bugs around the overlay and ripple behavior due to
react-transition-group modifying the DOM classname manually instead of
cloning a className prop into the child element.